requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
altgraph
Appium-Python-Client==2.0.0
asgiref==3.5.2
asn1crypto==1.5.1
async-generator==1.10
async-timeout==4.0.2
attrs==20.3.0
beautifulsoup4==4.11.1
bs4==0.0.1
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloud-sql-python-connector==0.9.0
colorama==0.4.6
cryptography==38.0.1
cs50==9.2.4
Cython==0.29.32
Django==4.1.2
exceptiongroup==1.0.0rc9
ez-setup==0.9
Flask==2.2.2
frida==16.0.8
frida-tools==12.0.4
frozenlist==1.3.1
fuzzywuzzy==0.18.0
google-api-core==2.10.2
google-api-python-client==2.64.0
google-auth==2.12.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.3
google-cloud==0.34.0
google-cloud-vision==3.1.4
googleapis-common-protos==1.56.4
grpcio==1.49.1
grpcio-status==1.49.1
gspread==5.6.0
h11==0.14.0
haversine==2.7.0
httplib2==0.20.4
idna==3.4
importlib-metadata==5.1.0
itsdangerous==2.1.2
jarowinkler==1.2.3
jellyfish==0.9.0
Jinja2==3.1.2
lib50==3.0.4
MarkupSafe==2.1.1
multidict==6.0.2
mysql-connector-python==8.0.31
numpy==1.23.3
oauth2client==4.1.3
oauthlib==3.2.1
outcome==1.2.0
pandas==1.5.0
pexpect==4.8.0
pg8000==1.29.2
Pillow==9.2.0
prompt-toolkit==3.0.36
proto-plus==1.22.1
protobuf==3.20.1
psycopg2-binary==2.9.4
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.14.0
pyobjc-core==8.5.1
pyobjc-framework-Cocoa==8.5.1
pyobjc-framework-SystemConfiguration==8.5.1
pyOpenSSL==22.1.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.4
PyYAML==5.4.1
requests==2.28.1
requests-oauthlib==1.3.1
requests-pkcs12==1.14
rsa==4.9
scramp==1.4.2
selenium==4.2.0
six==1.16.0
smsactivate==1.5
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.42
sqlparse==0.4.3
strsimpy==0.2.1
termcolor==1.1.0
transliterate==1.10.2
trio==0.22.0
trio-websocket==0.9.2
uritemplate==4.1.1
urllib3==1.26.12
urllib3-secure-extra==0.1.0
wcwidth==0.2.5
Werkzeug==2.2.2
wsproto==1.2.0
yarl==1.8.1
zipp==3.11.0
